+++ src/main/java/org/jabref/gui/groups/GroupDescriptions.java
@@ -1,13 +1,5 @@
 package org.jabref.gui.groups;
 
-import java.util.ArrayList;
-import java.util.regex.PatternSyntaxException;
-
-import javafx.scene.Node;
-import javafx.scene.text.Text;
-import javafx.scene.text.TextFlow;
-
-import org.jabref.gui.util.TooltipTextUtil;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.groups.ExplicitGroup;
 import org.jabref.model.groups.KeywordGroup;
@@ -17,29 +9,6 @@ import org.jabref.model.strings.StringUtil;
 public class GroupDescriptions {
 
     private GroupDescriptions() {
-    }
-
-    public static String getDescriptionForPreview(String field, String expr, boolean caseSensitive, boolean regExp) {
-        String header = regExp ? Localization.lang(
-                "This group contains entries whose <b>%0</b> field contains the regular expression <b>%1</b>",
-                field, expr) : Localization.lang(
-                        "This group contains entries whose <b>%0</b> field contains the keyword <b>%1</b>",
-                        field, expr);
-        String caseSensitiveText = caseSensitive ? Localization.lang("case sensitive") : Localization
-                .lang("case insensitive");
-        String footer = regExp ? Localization
-                .lang("Entries cannot be manually assigned to or removed from this group.") : Localization.lang(
-                        "Additionally, entries whose <b>%0</b> field does not contain "
-                                + "<b>%1</b> can be assigned manually to this group by selecting them "
-                                + "then using either drag and drop or the context menu. "
-                                + "This process adds the term <b>%1</b> to "
-                                + "each entry's <b>%0</b> field. "
-                                + "Entries can be removed manually from this group by selecting them "
-                                + "then using the context menu. "
-                                + "This process removes the term <b>%1</b> from "
-                                + "each entry's <b>%0</b> field.",
-                        field, expr);
-        return String.format("%s (%s). %s", header, caseSensitiveText, footer);
     }
 
     public static String getShortDescriptionKeywordGroup(KeywordGroup keywordGroup, boolean showDynamic) {
@@ -71,14 +40,6 @@ public class GroupDescriptions {
         }
         return sb.toString();
 
-    }
-
-    public static String getDescriptionForPreview() {
-        return Localization.lang("This group contains entries based on manual assignment. "
-                + "Entries can be assigned to this group by selecting them "
-                + "then using either drag and drop or the context menu. "
-                + "Entries can be removed from this group by selecting them "
-                + "then using the context menu.");
     }
 
     public static String getShortDescriptionExplicitGroup(ExplicitGroup explicitGroup) {
@@ -126,76 +87,4 @@ public class GroupDescriptions {
         }
         return sb.toString();
     }
-
-    protected static String fromTextFlowToHTMLString(TextFlow textFlow) {
-        StringBuilder htmlStringBuilder = new StringBuilder();
-        for (Node node : textFlow.getChildren()) {
-            if (node instanceof Text) {
-                htmlStringBuilder.append(TooltipTextUtil.textToHTMLString((Text) node));
-            }
-        }
-        return htmlStringBuilder.toString();
-    }
-
-    protected static String formatRegexException(String regExp, Exception e) {
-        String[] sa = e.getMessage().split("\\n");
-        StringBuilder sb = new StringBuilder();
-        for (int i = 0; i < sa.length; ++i) {
-            if (i > 0) {
-                sb.append("<br>");
-            }
-            sb.append(StringUtil.quoteForHTML(sa[i]));
-        }
-        String s = Localization.lang(
-                "The regular expression <b>%0</b> is invalid:",
-                StringUtil.quoteForHTML(regExp))
-                + "<p><tt>"
-                + sb
-                + "</tt>";
-        if (!(e instanceof PatternSyntaxException)) {
-            return s;
-        }
-        int lastNewline = s.lastIndexOf("<br>");
-        int hat = s.lastIndexOf('^');
-        if ((lastNewline >= 0) && (hat >= 0) && (hat > lastNewline)) {
-            return s.substring(0, lastNewline + 4) + s.substring(lastNewline + 4).replace(" ", "&nbsp;");
-        }
-        return s;
-    }
-
-    protected static ArrayList<Node> createFormattedDescription(String descriptionHTML) {
-        ArrayList<Node> nodes = new ArrayList<>();
-
-        descriptionHTML = descriptionHTML.replaceAll("<p>|<br>", "\n");
-
-        String[] boldSplit = descriptionHTML.split("(?=<b>)|(?<=</b>)|(?=<i>)|(?<=</i>)|(?=<tt>)|(?<=</tt>)|(?=<kbd>)|(?<=</kbd>)");
-
-        for (String bs : boldSplit) {
-
-            if (bs.matches("<b>[^<>]*</b>")) {
-
-                bs = bs.replaceAll("<b>|</b>", "");
-                Text textElement = new Text(bs);
-                textElement.setStyle("-fx-font-weight: bold");
-                nodes.add(textElement);
-            } else if (bs.matches("<i>[^<>]*</i>")) {
-
-                bs = bs.replaceAll("<i>|</i>", "");
-                Text textElement = new Text(bs);
-                textElement.setStyle("-fx-font-style: italic");
-                nodes.add(textElement);
-            } else if (bs.matches("<tt>[^<>]*</tt>|<kbd>[^<>]*</kbd>")) {
-
-                bs = bs.replaceAll("<tt>|</tt>|<kbd>|</kbd>", "");
-                Text textElement = new Text(bs);
-                textElement.setStyle("-fx-font-family: 'Courier New', Courier, monospace");
-                nodes.add(textElement);
-            } else {
-                nodes.add(new Text(bs));
-            }
-        }
-
-        return nodes;
-    }
-
 }